clean up exception repo file :
